mostfa's bidan error

+++ CIE205-Project-F2018-std/DSProject_Code_S2018/Castle/Castle.cpp
@@ -34,11 +34,11 @@ void Castle::tower_act(int tower_number, Enemy* enemy)
 	int distance = enemy->GetDistance();
 	double tower_power = Towers[index].get_power();
 	int type = enemy->get_type();
-	int k = 0;
+	double k = 0;
 	if (type == 1 || type == 3) k = 1;
 	else k = 0.5;
 	
-	double health_deducted =(1/distance)*(tower_power)*(1/k);
+	double health_deducted =(1.0/distance)*(tower_power)*(1/k);
 
 
 	enemy->set_health((enemy->get_health()) - health_deducted);   // deduct some health of the enemy;
